+++ internal/provider/provider.go
@@ -159,6 +159,7 @@ func (p *catoProvider) DataSources(_ context.Context) []func() datasource.DataSo
 	return []func() datasource.DataSource{
 		NewAccountSnapshotSiteDataSource,
 		AllocatedIpDataSource,
+		DhcpRelayDataSource,
 		LicensingInfoDataSource,
 		NetworkInterfacesDataSource,
 		SiteLocationDataSource,