+++ internal/provider/resource_wan_fw_rule.go
@@ -91,13 +91,12 @@ func (r *wanFwRuleResource) Schema(_ context.Context, _ resource.SchemaRequest, 
 						Optional:    true,
 					},
 					"index": schema.Int64Attribute{
-						Description: "Ruile index",
-						Required:    false,
-						Optional:    true,
+						Description: "Rule Index",
+						Computed:    true,
+						Optional:    false,
 						PlanModifiers: []planmodifier.Int64{
-							int64planmodifier.UseStateForUnknown(), // Avoid drift
+							int64planmodifier.UseStateForUnknown(),
 						},
-						Computed: true,
 					},
 					"enabled": schema.BoolAttribute{
 						Description: "Attribute to define rule status (enabled or disabled)",
@@ -1349,9 +1348,13 @@ func (r *wanFwRuleResource) Schema(_ context.Context, _ resource.SchemaRequest, 
 						},
 					},
 					"service": schema.SingleNestedAttribute{
-						Description: "Destination service traffic matching criteria. Logical ‘OR’ is applied within the criteria set. Logical ‘AND’ is applied between criteria sets.",
+						Description: "Destination service traffic matching criteria. Logical 'OR' is applied within the criteria set. Logical 'AND' is applied between criteria sets.",
 						Required:    false,
 						Optional:    true,
+						PlanModifiers: []planmodifier.Object{
+							objectplanmodifier.UseStateForUnknown(), // Avoid drift
+						},
+						Computed: true,
 						Attributes: map[string]schema.Attribute{
 							"standard": schema.SetNestedAttribute{
 								Description: "Standard Service to which this Internet Firewall rule applies",
@@ -1662,6 +1665,7 @@ func (r *wanFwRuleResource) Schema(_ context.Context, _ resource.SchemaRequest, 
 						Description: "The set of exceptions for the rule. Exceptions define when the rule will be ignored and the firewall evaluation will continue with the lower priority rules.",
 						Required:    false,
 						Optional:    true,
+						Computed:    true,
 						PlanModifiers: []planmodifier.Set{
 							setplanmodifier.UseStateForUnknown(), // Avoid drift
 						},